requirements.txt
pyTelegramBotAPI==4.14.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
python-json-logger==2.0.7
gunicorn==21.2.0
PyJWT==2.8.0
